강원신문 / 전자신문 동시 자동 크롤링

+++ src/main/java/com/example/crawling/vo/Book.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Document(collection = "crawling")  /*MongoDB의 test 컬렉션과 연결하도록 설정*/
 public class Book {
     @Id
-    private int newsNo;
+    private long newsNo;
     private String url;    // URL 타입?
     private String title;
     private String date;
